[JobbyServer] modify output of getUser success response.

+++ src/main/kotlin/com/bravedroid/jobby/auth/infrastructure/repositories/UserJpaRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UserJpaRepository : JpaRepository<UserEntity, Int> {
+interface UserJpaRepository : JpaRepository<UserEntity, Long> {
     fun findByEmail(email: String): UserEntity?
 }
